Fix net store race condition
Arguments of the `Net:PostHTTPAsync` function appear to be evaluated asynchronously in addition to the function itself. Therefore, the `playerRankObject` may no longer exist  when it is evaluated due to the object being deleted shortly after, leading to a nil value index attempt. This commit fixes this issue by first ensuring the GUID passed to the store function is not nil, and then performs a closure over it to a local value to prevent async race conditions.