Fix/Mitigate Nil Player GUIDs on Net Store
Added `r_PlayerGuidStr` instance member to PlayerRank class to hold a reliable string copy of the player's GUID. This is a workaround to the unreliability of `r_Player.guid` being `nil` intermittently when trying to net store player progression. This also has the benefit of cleaning up the many `ToString` calls around the codebase. Lastly, did some refactoring around `PlayerXPUpdated`.