[Feat] Add Global Progression (#16)
* Add JSON Library & Config Params.

* [Refactor] Storage Management

Refactored the existing local SQLite storage system to work within the new StorageManager system. Also preps boilerplate for net storage.

* Added callback structure & net storage auth

- Net storage calls will be async, so a callback structure for fetching player data is required.
- Added net storage initialization, which occurs on first boot and mod reloads. This checks if the mod can reach the Global Progression server, if the mod is too outdated to communicate, and if the mod is authenticated.
- Added mod version print on load.
- Made `AddNewPlayerForStats` a local-only event and removed its unused `data` parameter.
- Fixed players getting deregistered from `currentRankupPlayers` upon mod reload (stats up to that point will be lost though; this is primarily for testing).

* Add Net Storage fetch functionality

* Prevent XP multiplier

Prevent XP multiplier when Global Progression is enabled.

* Add Net Storage store functionality

* Update API endpoints & README and bug fixes

- Updated endpoints for API v0.1.3.
- Global player data will be loaded over local data if the global total score is exactly one. This is a half-baked approach to add player progression resets, but progression will be fast-forwarded again if the player earns score and then joins another server with higher pre-recorded score.
- Added the ability for the Global Progression master server to set the XP multiplier upon authentication.
- Fixed a bug with player data mapping that would cause the name or GUID of a playerRankObject to become `null`.
- Fixed a bug with error reporting in `NetStorage:storePlayerProgress`.
- Updated README to include info on Global Progression, and improved verbiage for installation, configuration, and known issues.

* Added Round handling for storage & fixed bugs

- Added handling & storage of Round data. New functions made available in StorageManager, but only NetStorage is currently affected by calling these functions.
- Refactored PlayerRank to store the whole Player object to avoid extra GUID lookup calls.
- Removed unneeded JSON library, because it's actually built into VEXT.
- Fixed a bug where new players could not be properly submitted to the Global Progression server.
- Fixed closure issues in async callback functions.
- Fixed a bug where the XP Multiplier could turn XP from an integer to a float.
- Fixed a bug where respawning from the menu ("Suicide") earned a kill against yourself.
- Other light refactoring.

* Trim trailing slash from GlobalProgression URL

Ensures `CONFIG.GlobalProgression.url` does not end with a slash by removing it if present upon init.

* Fix net store race condition

Arguments of the `Net:PostHTTPAsync` function appear to be evaluated asynchronously in addition to the function itself. Therefore, the `playerRankObject` may no longer exist  when it is evaluated due to the object being deleted shortly after, leading to a nil value index attempt. This commit fixes this issue by first ensuring the GUID passed to the store function is not nil, and then performs a closure over it to a local value to prevent async race conditions.

* Convert fetched net data ambiguous floats to ints

Fixes `!levels` chat command showing decimal values.

* Fixed round finalization including bots in player count

Bots are no longer included in player count when finalizing/reporting a round.

* Fix TIP formatting in README.md

* Fix link to Releases page in README

* Fix/Mitigate Nil Player GUIDs on Net Store

Added `r_PlayerGuidStr` instance member to PlayerRank class to hold a reliable string copy of the player's GUID. This is a workaround to the unreliability of `r_Player.guid` being `nil` intermittently when trying to net store player progression. This also has the benefit of cleaning up the many `ToString` calls around the codebase. Lastly, did some refactoring around `PlayerXPUpdated`.